스플래시 액티비티 완성
